# shared-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Sereno API
  version: "1.0.0"
  description: >
    API do piloto de neuromodulação não invasiva. Contrato é a fonte de verdade.
    Erros seguem problem+json (RFC 9457). O braço do estudo nunca é exposto:
    endpoints de participante recebem apenas um handle de protocolo neutro, e a
    exportação de pesquisa usa braço codificado (A/B).
servers:
  - url: https://api.sereno.example/v1
security:
  - bearerAuth: []
paths:
...
  /participants/me/status:
    get:
      summary: Andamento do participante no protocolo (semana, adesão, T2)
      operationId: getMyProgress
      description: >-
        Em que semana o participante está, quantas sessões contaram para a adesão, se a
        avaliação intermediária (T2, 2ª semana) está aberta e se houve descontinuação de
        protocolo. É o que permite convidar para o T2 **na hora certa**. Aplica de passagem a
        regra de adesão da 2ª semana. NÃO revela braço, condição nem escore. Participante
        (`progress:read`).
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ParticipantProgress' } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /participants/{id}/discontinue:
    post:
      summary: Registrar descontinuação de protocolo (staff)
      operationId: discontinueParticipant
      description: >-
        Os dois critérios de descontinuação que são juízo humano: pedido do participante e
        evento adverso que contraindique a continuidade. O terceiro (adesão < 50% ao final da
        2ª semana) é regra e o servidor a aplica sozinho. Quem descontinua **permanece na
        análise por intenção de tratar** — é o que separa este estado da retirada de
        consentimento. Staff de inscrição (`enroll:write`).
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/DiscontinueIn' } } }
      responses:
        '201': { description: Registrada, content: { application/json: { schema: { $ref: '#/components/schemas/Discontinuation' } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /discontinuations:
    get:
      summary: Listar descontinuações de protocolo (equipe)
      operationId: listDiscontinuations
      description: >-
        Saídas do protocolo, pseudonimizadas (código do estudo, motivo, contagem que motivou a
        decisão). Alimenta o fluxo de participantes do CONSORT e o relatório parcial ao CEP.
        Nunca traz braço nem PII.
      parameters:
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/DiscontinuationPage' } } } }
        '403': { $ref: '#/components/responses/Problem' }
  /discontinuations/evaluate:
    post:
      summary: Aplicar a regra de adesão da 2ª semana a todos os ativos (staff)
      operationId: evaluateDiscontinuations
      description: >-
        Varredura idempotente: descontinua quem, ao final da 2ª semana, concluiu menos de 50%
        das sessões previstas até ali. Existe porque a avaliação preguiçosa (no início de
        sessão e na tela inicial) nunca alcança quem parou de abrir o aplicativo — que é
        exatamente o caso que a regra precisa pegar. Staff de inscrição (`enroll:write`).
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/DiscontinuationSweep' } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
...
  /allocation:
    post:
      summary: Alocar participante (staff) — resposta NEUTRA, nunca revela o braço
      operationId: allocateParticipant
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AllocateIn' } } }
      responses:
        '201': { description: Alocado (confirmação neutra), content: { application/json: { schema: { $ref: '#/components/schemas/AllocationAccepted' } } } }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /allocation/{id}/unblind-request:
    post:
      summary: Solicitar desbloqueio de um participante (admin) — passo 1 de 2, NÃO revela
      operationId: unblindRequest
      description: >
        Passo 1 do desbloqueio em DUAS PESSOAS: um admin (`unblind:request`) abre um pedido
        justificado. NÃO revela a condição — apenas registra a pendência e audita o pedido
        (`unblind.requested`, sem a condição). A revelação só ocorre no passo 2 (`unblind-approve`),
        exigido a um SEGUNDO admin distinto. 409 se já houver pedido pendente ou já desbloqueado.
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UnblindIn' } } }
      responses:
        '200': { description: Pedido registrado (aguarda 2º aprovador), content: { application/json: { schema: { $ref: '#/components/schemas/UnblindRequestResult' } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /allocation/{id}/unblind-approve:
    post:
      summary: Aprovar desbloqueio (SEGUNDO admin) — passo 2 de 2, revela a condição
      operationId: unblindApprove
      description: >
        Passo 2 do desbloqueio em duas pessoas: um SEGUNDO admin (`unblind:request`), DISTINTO do
        solicitante, aprova o pedido pendente. Só aqui a condição (ativo/sham) é revelada, usando a
        chave selada; a auditoria registra `unblind.performed` (quem aprovou/quando) SEM a condição
        em claro. 409 se não houver pedido pendente, se o aprovador for o mesmo solicitante, ou se já
        desbloqueado. É o ÚNICO caminho da API para a condição.
      parameters: [ { $ref: '#/components/parameters/PathId' } ]
      responses:
        '200': { description: Condição revelada (ao 2º admin), content: { application/json: { schema: { $ref: '#/components/schemas/UnblindResult' } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    PathId: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    Limit: { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
  responses:
    Problem:
      description: Erro (application/problem+json, RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
      required: [title, status]
...
    AllocateIn:
      type: object
      properties:
        participant_id: { type: string, format: uuid }
      required: [participant_id]
...
    ParticipantProgress:
      type: object
      description: >-
        Andamento no protocolo. Idêntico nos dois braços — nada aqui depende da condição.
      properties:
        status: { type: string, enum: [active, withdrawn, completed, removed, discontinued] }
        allocated: { type: boolean }
        study_day: { type: [integer, "null"], description: "dia de estudo (1 = dia da alocação)" }
        study_week: { type: [integer, "null"] }
        sessions_completed: { type: integer, description: "sessões com pelo menos 80% da duração" }
        sessions_prescribed: { type: integer }
        adherence_pct: { type: number }
        t2: { oneOf: [ { $ref: '#/components/schemas/T2Window' }, { type: "null" } ] }
        discontinuation: { oneOf: [ { $ref: '#/components/schemas/DiscontinuationBrief' }, { type: "null" } ] }
      required: [status, allocated, sessions_completed, sessions_prescribed, adherence_pct]
    T2Window:
      type: object
      description: "Avaliação intermediária da 2ª semana: quando abre, até quando, e se já foi feita."
      properties:
        opens_at: { type: string, format: date-time }
        closes_at: { type: string, format: date-time }
        due: { type: boolean, description: "janela aberta e ainda não respondida" }
        late: { type: boolean, description: "janela fechada sem resposta" }
        completed: { type: boolean }
      required: [opens_at, closes_at, due, late, completed]
    DiscontinuationBrief:
      type: object
      properties:
        reason: { type: string, enum: [solicitacao_participante, evento_adverso, adesao_insuficiente] }
        decided_at: { type: string, format: date-time }
        kept_in_itt: { type: boolean, description: "sempre true — descontinuar não exclui da análise" }
      required: [reason, decided_at, kept_in_itt]
    DiscontinueIn:
      type: object
      description: >-
        Só os motivos que são juízo humano; `adesao_insuficiente` é regra do servidor.
      properties:
        reason: { type: string, enum: [solicitacao_participante, evento_adverso] }
        adverse_event_id: { type: [string, "null"], format: uuid }
      required: [reason]
    Discontinuation:
      type: object
      properties:
        id: { type: string, format: uuid }
        study_code: { type: string, description: "pseudônimo; nunca PII" }
        reason: { type: string, enum: [solicitacao_participante, evento_adverso, adesao_insuficiente] }
        study_week: { type: [integer, "null"] }
        sessions_completed: { type: [integer, "null"] }
        sessions_prescribed: { type: [integer, "null"] }
        kept_in_itt: { type: boolean }
        decided_at: { type: string, format: date-time }
      required: [id, study_code, reason, kept_in_itt, decided_at]
    DiscontinuationPage:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Discontinuation' } }
      required: [items]
    DiscontinuationSweep:
      type: object
      properties:
        evaluated_at: { type: string, format: date-time }
        discontinued: { type: integer, description: "quantos a varredura descontinuou agora" }
      required: [evaluated_at, discontinued]
...
    AllocationAccepted:
      type: object
      description: "Confirmação neutra — deliberadamente SEM o braço (ocultação da alocação)."
      properties:
        status: { type: string, enum: [allocated] }
        block: { type: integer }
      required: [status, block]
    UnblindIn:
      type: object
      description: "Justificativa obrigatória do desbloqueio (auditada; nunca contém a condição)."
      properties:
        justification: { type: string, minLength: 10, maxLength: 500 }
      required: [justification]
    UnblindRequestResult:
      type: object
      description: "Passo 1: pedido registrado; NÃO contém a condição (aguarda 2º aprovador)."
      properties:
        participant_id: { type: string, format: uuid }
        status: { type: string, enum: [pending_approval] }
        requested_at: { type: string, format: date-time }
      required: [participant_id, status, requested_at]
    UnblindResult:
      type: object
      description: "Passo 2: condição revelada ao 2º admin — ÚNICO ponto onde a condição aparece na API."
      properties:
        participant_id: { type: string, format: uuid }
        condition: { type: string, enum: [active, sham] }
        unblinded_at: { type: string, format: date-time }
        justification: { type: string }
      required: [participant_id, condition, unblinded_at]
...